Moved custom hooks into seperate files and removed audio loadded dependency
